[WIP][PROOF OF CONCEPT DO NOT MERGE TO MASTER] Filter the GVKs to desired ones.
DO NOT MERGE INTO THE DEFINITIVE BRANCH UNTIL EVERYTHING IS REVIEWED
AND CLARIFIED

WARNING at the moment the only changes added have been the necessary ones
to make the operator not return error logs.
Before this commit the operator was returning errors due to
the operator-sdk implementation of the serveCRMetrics function was
trying to serve all GVKs returned by the used Schema by the operator.
This included objects that are not directly used in the operator code
and thus the operator does not have role permissions to use them.
By filtering only the schemas that is avoided but the current implementation
manually writes the Kinds, Groups and has some filtering limitations (see
code comments) and the filtering implementation could be considered
somewhat buggy.
Also it is not clear at the moment what's the purpose of those metrics
and which ones should be exposed.